Move AuthContext up to _app, move dash pages into /admin folder
