+++ components/DashboardLayout/Sidebar.tsx
@@ -1,11 +1,7 @@
-import { User } from "@prisma/client";
 import { useRouter } from "next/router";
 import Link from "next/link";
-import React from "react";
-
-type Props = {
-  user: User | null;
-};
+import React, { useContext } from "react";
+import { AuthContext } from "pages/_app";
 
 type SidebarLinkProps = {
   href: string;
@@ -30,24 +26,25 @@ const SidebarLink: React.FC<SidebarLinkProps> = ({
   );
 };
 
-const Sidebar: React.FC<Props> = ({ user }: Props) => {
+const Sidebar: React.FC = () => {
+  const user = useContext(AuthContext);
   return (
-    <div className="fixed flex flex-col justify-between w-50 h-screen border-r border-gray-500">
+    <div className="fixed top-0 flex flex-col justify-between w-56 h-screen border-r border-gray-500">
       <div className="px-12 py-24">
         <div className="w-full flex justify-center mb-16">
           <div className="w-20 h-20 bg-gray-500 rounded-full" />
         </div>
-        <SidebarLink href="/">Participants</SidebarLink>
-        <SidebarLink href="/f">All Users</SidebarLink>
-        <SidebarLink href="/f">Invite Users</SidebarLink>
-        <SidebarLink href="/f">Calendar</SidebarLink>
-        <SidebarLink href="/f">Notes</SidebarLink>
+        <SidebarLink href="/admin/participants">Participants</SidebarLink>
+        <SidebarLink href="/admin/users">All Users</SidebarLink>
+        <SidebarLink href="/admin/invite">Invite Users</SidebarLink>
+        <SidebarLink href="/admin/calendar">Calendar</SidebarLink>
+        <SidebarLink href="/admin/notes">Notes</SidebarLink>
       </div>
       <div className="mb-12 px-6">
         <div className="flex items-center">
           <div className="w-10 h-10 mr-4 bg-gray-500 rounded-full" />
           <div>
-            <p>{user?.name}</p>
+            <p className="font-medium">{user?.name}</p>
             <p className="text-gray-500">Admin</p>
           </div>
         </div>
