+++ pages/admin/invite.tsx
@@ -1,0 +1,14 @@
+import DashboardLayout from "components/DashboardLayout";
+import React from "react";
+
+const AdminInvitePage: React.FC = () => {
+  return (
+    <DashboardLayout>
+      <div className="mx-16 mt-24">
+        <h1 className="text-3xl font-display font-medium">Invite Users</h1>
+      </div>
+    </DashboardLayout>
+  );
+};
+
+export default AdminInvitePage;
